# plataforma-recursos/services/interface/swagger.yaml
openapi: 3.0.3
info:
  title: EW2026 Interface Service
  version: 1.0.0
  description: Interface web server-side (Pug) da plataforma de recursos educativos.
servers:
  - url: http://localhost:16020
    description: Interface via gateway
  - url: http://localhost:16026
    description: Interface direta
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ew.sid
...
paths:
...
  /posts:
    get:
      tags: [Posts]
      summary: Listar posts
      responses:
        '200':
          description: HTML com lista de posts
          content:
            text/html:
              schema:
                type: string
    post:
      tags: [Posts]
      summary: Criar novo post
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [titulo, conteudo]
              properties:
                titulo:
                  type: string
                conteudo:
                  type: string
                resourceId:
                  type: string
      responses:
        '302':
          description: Redireciona para detalhe do post ou para formulário em erro
  /posts/new:
    get:
      tags: [Posts]
      summary: Formulário para criar post
      security:
        - sessionCookie: []
      responses:
        '200':
          description: HTML do formulário
          content:
            text/html:
              schema:
                type: string
        '302':
          description: Redirecionamento caso sem permissões
  /posts/{id}:
    get:
      tags: [Posts]
      summary: Ver detalhe de post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: HTML com detalhe do post
          content:
            text/html:
              schema:
                type: string
  /posts/{id}/edit:
    get:
      tags: [Posts]
      summary: Formulário de edição de post
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: HTML do formulário de edição
          content:
            text/html:
              schema:
                type: string
        '302':
          description: Redirecionamento em falta de permissões
    post:
      tags: [Posts]
      summary: Guardar alterações de post
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [titulo, conteudo]
              properties:
                titulo:
                  type: string
                conteudo:
                  type: string
                resourceId:
                  type: string
      responses:
        '302':
          description: Redireciona para o detalhe do post
  /posts/{id}/comments:
    post:
      tags: [Posts]
      summary: Criar comentário num post
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [texto]
              properties:
                texto:
                  type: string
      responses:
        '302':
          description: Redireciona para o detalhe do post
  /posts/{id}/comments/{cid}/delete:
    post:
      tags: [Posts]
      summary: Remover comentário de post
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: cid
          in: path
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redireciona para o detalhe do post
  /posts/{id}/delete:
    post:
      tags: [Posts]
      summary: Remover post
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redireciona para lista de posts
  /admin/users:
    get:
      tags: [Admin]
      summary: Listar utilizadores
      security:
        - sessionCookie: []
      responses:
        '200':
          description: HTML com gestão de utilizadores
          content:
            text/html:
              schema:
                type: string
  /admin/users/{id}/role:
    post:
      tags: [Admin]
      summary: Atualizar role de utilizador
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [admin, produtor, consumidor]
      responses:
        '302':
          description: Redireciona para gestão de utilizadores
  /admin/users/{id}/delete:
    post:
      tags: [Admin]
      summary: Remover utilizador
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redireciona para gestão de utilizadores
  /admin/news:
    get:
      tags: [Admin]
      summary: Listar notícias
      security:
        - sessionCookie: []
      responses:
        '200':
          description: HTML com gestão de notícias
          content:
            text/html:
              schema:
                type: string
    post:
      tags: [Admin]
      summary: Criar notícia
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [titulo, conteudo]
              properties:
                titulo:
                  type: string
                conteudo:
                  type: string
      responses:
        '302':
          description: Redireciona para gestão de notícias
  /admin/news/{id}/delete:
    post:
      tags: [Admin]
      summary: Remover notícia
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redireciona para gestão de notícias
  /admin/export:
    get:
      tags: [Admin]
      summary: Exportar dump de dados
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Ficheiro JSON para download
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '302':
          description: Redireciona para dashboard de dados em erro
  /admin/import:
    get:
      tags: [Admin]
      summary: Formulário de importação de dump JSON
      security:
        - sessionCookie: []
      responses:
        '200':
          description: HTML do formulário de importação
          content:
            text/html:
              schema:
                type: string
    post:
      tags: [Admin]
      summary: Importar dump JSON
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [dump]
              properties:
                dump:
                  type: string
                  description: JSON completo no textarea
      responses:
        '200':
          description: HTML com relatório de importação
          content:
            text/html:
              schema:
                type: string
        '302':
          description: Redireciona para o formulário em caso de erro
  /admin/data:
    get:
      tags: [Admin]
      summary: Dashboard de operações de dados
      security:
        - sessionCookie: []
      responses:
        '200':
          description: HTML do dashboard de dados
          content:
            text/html:
              schema:
                type: string